Add files via upload
DO NOT RUN

Documentation for rough draft
(pseudocode and code)

program includes file manipulation due to login and score info changed locally (as of right now).

--------------------------------------------
first you will encounter a log-in screen
you will be given the option to login, sign-up or quit

if you login, the program will search for a file called users.txt
if found, it will search for your user and password in a list

["Nsj", "1001"]

if found, you will log in

(after playing it will append the list with your new score)

["Nsj", "1001", "100"]

it will also take this list, ignore [0] and [1] which are your user and password, and rearrange the rest to put your highest score in order. [3] will be the highest score.
------------------------------------------

New file

I've created a new file that has three lists, easy_hangman, medium_hangman and hard_hangman. Each file has 50 words in it corresponding to the level of difficulty. 

During the main program execution, you will have an option to select game mode right after logging in. Depending on game mode you choose, you will enter a loop that will end once you guess the word or reach 6. every wrong answer will print a hangman image corresponding to the amount of wrong answers. Once you reach 6, you will be put in the GAMEOVER function that will end the game and populate the last image of the full hangman.

correct_guesses will be calculated as well, 25 for each letter in easy, 50 for each letter in medium, 100 for each letter in hard.

# not coded yet: I want to put in a way for you to replay the game as well after you correctly guess the word, if that makes sense. if you don't the game 0ver function will save your score and exit the program.

Game over

if you reach 6 guesses or fill up the hang man

        # print score
        # pull users.txt
        # check index 3 and up for scores
        # rearrange so highest is in index 3
        # print all scores

-----------------------------------

Call main, after the program finishes the login screen, it will not replay. (created this so that we can work on a replay element).